// ui/src/generated/Property.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PropValue } from "./PropValue";

export type Property = { 
/**
 * The key; nested maps are flattened to dotted keys such as `metadata.type`.
 */
key: string, value: PropValue, };